+++ CHANGELOG.md
@@ -18,11 +18,36 @@ export const myActionsWithGate = wrapActionsWithReducerIdGate(reducerId, {
 
 It returns the same actions, but wrapped with reducerIdGate action.
 
+### Action Menu
+
+- All types of `ActionMenuButton` support keyboard navigation, whether `portalTarget` is used or not
+- `ActionMenuItem` has new props `id` and `loading`
+- `ActionMenuItem` receives the HTML `disabled` prop when disabled
+- `ActionMenuLink` can be triggered with Space, just like `ActionMenuItem`
+- `ActionMenuLink` has border radius if first or last child, just like `ActionMenuItem`
+- The text in `ActionMenuItem` and `ActionMenuLink` does not wrap
+
+### StandardTable
+
+- Sort order icon `amount` variant now has inverted arrow direction.
+- When using editable cells, `editorValue` was not always reset and could
+become inconsistent with the value of the data.
+`editorValue` is now always reset when editor for a cell is opened.
+- `rowBackgroundResolver` can now return color or object containing `background` and `hoverBackground` fields.
+
+### RadioButton
+
+- `RadioButton` has fixed width, i.e. will not be squashed when content is competing for space
+
+### Dependencies
+
+Font awesome icons have been updated to `5.15.3`.
+
 ## 10.1.2
 
 ### StandardTable
 
-- `rowBackgroundResolver` now gets `selected` as second argument.
+- `rowBackgroundResolver` function now gets `selected` as second argument.
   This makes it possible to change background color depending on
   the checkbox being checked or not.